continue IDEA-72758, init stable support for embed statements in user mxml files (server side done, client side is not yet) IDEA-72629

+++ flex/tools/flex-ui-designer/idea-plugin/src/com/intellij/flex/uiDesigner/abc/MovieTranscoder.java
@@ -85,10 +85,10 @@ public class MovieTranscoder extends SwfTranscoder {
       if (writeBounds) {
         buffer.order(ByteOrder.BIG_ENDIAN);
         buffer.position(0);
-        buffer.putShort((short)bounds.x);
-        buffer.putShort((short)bounds.y);
-        buffer.putShort((short)bounds.width);
-        buffer.putShort((short)bounds.height);
+        buffer.putInt(bounds.x);
+        buffer.putInt(bounds.y);
+        buffer.putInt(bounds.width);
+        buffer.putInt(bounds.height);
         out.write(buffer.array(), 0, buffer.position());
         buffer.order(ByteOrder.LITTLE_ENDIAN);
       }
